+++ R/soread.R
@@ -13,9 +13,17 @@
 #'your workspace called "\code{mydf}" unless specified using the \code{out}
 #'argument.
 #'
-#' @param sep character. Determines the field separator character passed to \code{read.table}.
-#' @param header logical. Determines whether the first row consists of names of variables.
-#' @param stringsAsFactors logical. Whether strings are converted to factors or remain character variables.
+#' @param sep character. Determines the field separator character passed 
+#' to \code{read.table}.
+#' @param header logical. Determines whether the first row consists of 
+#' names of variables.
+#' @param stringsAsFactors logical. Whether strings are converted to 
+#' factors or remain character variables. Defaults to `FALSE`
+#' @param skipAfterHeader numeric. Some newer print methods print additional
+#' information just below the headers (for example, `tibble`s and `data.table`s).
+#' If `skipAfterHeader = TRUE`, the second line will be removed. If 
+#' `skipAfterHeader` is a numeric value, those lines after the header row
+#' will be removed before reading the data.
 #' @param out character. Desired output object name. Defaults to \code{"mydf"}.
 #' @return A data.frame as \code{read.table} produces.
 #' @note By default, stringsAsFactors is FALSE which is different to the R default.
@@ -38,10 +46,18 @@
 #'soread()
 #'}
 #'
-soread <- function(sep = "", header = TRUE, stringsAsFactors = FALSE, out = "mydf") {
+soread <- function(sep = "", header = TRUE, stringsAsFactors = FALSE, skipAfterHeader = NULL, out = "mydf") {
   temp <- sub("^[# ]+", "", suppressWarnings(readClip()))
+  if (!is.null(skipAfterHeader)) {
+    if (is.logical(skipAfterHeader) & isTRUE(skipAfterHeader)) {
+      temp <- temp[-2]
+    } else {
+      temp <- temp[-c(1 + skipAfterHeader)]
+    } 
+  } 
   temp <- read.table(text = temp, header = header, stringsAsFactors = stringsAsFactors, sep = sep)
   assign(out, temp, envir = .GlobalEnv)
   message("data.frame ", dQuote(out), " created in your workspace")
   temp
 }
+NULL